+++ packages/back-to-top/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @italia/back-to-top
 
+## 1.0.0-alpha.17
+
+### Patch Changes
+
+- Aggiornamento della versione
+- Updated dependencies []:
+  - @italia/globals@1.0.0-alpha.17
+
 ## 1.0.0-alpha.16
 
 ### Patch Changes