+++ packages/carousel/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @italia/carousel
 
+## 1.0.0-alpha.17
+
+### Patch Changes
+
+- Aggiornamento della versione
+- Updated dependencies []:
+  - @italia/card@1.0.0-alpha.17
+  - @italia/globals@1.0.0-alpha.17
+  - @italia/i18n@1.0.0-alpha.17
+
 ## 1.0.0-alpha.16
 
 ### Patch Changes