+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @italia/tabs
 
+## 1.0.0-alpha.17
+
+### Patch Changes
+
+- Aggiornamento della versione
+- Updated dependencies []:
+  - @italia/globals@1.0.0-alpha.17
+  - @italia/i18n@1.0.0-alpha.17
+
 ## 1.0.0-alpha.16
 
 ### Patch Changes