+++ packages/video/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @italia/video
 
+## 1.0.0-alpha.17
+
+### Patch Changes
+
+- Aggiornamento della versione
+- Updated dependencies []:
+  - @italia/button@1.0.0-alpha.17
+  - @italia/checkbox@1.0.0-alpha.17
+  - @italia/globals@1.0.0-alpha.17
+  - @italia/i18n@1.0.0-alpha.17
+  - @italia/icon@1.0.0-alpha.17
+
 ## 1.0.0-alpha.16
 
 ### Patch Changes